Sketch out edges
