+++ packages/entities/src/index.js
@@ -27,3 +27,4 @@ export { timestamp, toDate } from './timestamp';
 
 export { VizHubAPIError } from './VizHubAPIError';
 export { Commit } from './commit';
+export { Edge } from './edge';
